requirements.txt
django
django-rest-framework
openai
django-environ
pytest
pytest-django
pytest-mock
drf-spectacular